backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
chromadb==0.5.20
sentence-transformers==3.0.1
google-generativeai==0.7.2
numpy
pandas
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1
websockets==12.0
faker==24.11.0
pytest==8.2.0
pytest-asyncio==0.23.6
